Move DynamicSurfaceAberration class.
